+++ src/main/java/moze_intel/projecte/gameObjs/gui/GUICondenserMK2.java
@@ -49,7 +49,7 @@ public class GUICondenserMK2 extends GuiContainer
 	@Override
 	protected void drawGuiContainerForegroundLayer(int var1, int var2)
 	{
-		int toDisplay = container.displayEmc > container.requiredEmc ? container.requiredEmc : container.displayEmc;
+		long toDisplay = container.displayEmc > container.requiredEmc ? container.requiredEmc : container.displayEmc;
 		this.fontRenderer.drawString(Constants.EMC_FORMATTER.format(toDisplay), 140, 10, 4210752);
 	}
 }
