+++ src/main/java/moze_intel/projecte/utils/Constants.java
@@ -13,7 +13,7 @@ public final class Constants
 	public static final float[] EXPLOSIVE_LENS_RADIUS = new float[] {4.0F, 8.0F, 12.0F, 16.0F, 16.0F, 16.0F, 16.0F, 16.0F};
 	public static final int[] EXPLOSIVE_LENS_COST = new int[] {384, 768, 1536, 2304, 2304, 2304, 2304, 2304};
 	
-	public static final int TILE_MAX_EMC = Integer.MAX_VALUE;
+	public static final double TILE_MAX_EMC = Double.MAX_VALUE;
 	
 	public static final int COLLECTOR_MK1_MAX = 10000;
 	public static final int COLLECTOR_MK2_MAX = 30000;
